requirements.txt
livekit>=0.11.0
livekit-agents>=0.11.0
livekit-plugins-openai>=0.11.0
livekit-plugins-elevenlabs>=0.11.0
livekit-plugins-silero>=0.11.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

